pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[project]
name = "image2base64"
version = "1.0.0"
dependencies = [
    "pillow",
    "pyperclip"
]

[project.scripts]
image2base64 = "image2base64.main:main"